Move FCoeff file generation to separate module
